refactor: 迁移 javascript 代码
