+++ app/javascript/controllers/admin/index.js
@@ -1,39 +1,7 @@
-import compareVersion from 'compare-versions';
-const newVesionUrl = 'https://api.github.com/repos/tryzealot/zealot/releases/latest';
+import { zealotNewVersionEventer } from "./new_version"
+import { addRestartSericeEventer } from "./restart_service"
 
-function insert_link(elm, title, link) {
-  elm.html(
-    '<a target="_blank" href="' + link + '">' +
-    '<i class="fas fa-rainbow"></i>' + title + '</a>');
-}
-
-function check_new_version() {
-  var elm = $('#new-version');
-  if (!elm.length) { return; }
-
-  var current_version = $('#current-version').html();
-  var title = elm.data('title');
-  if (current_version == 'development') {
-    insert_link(elm, title, 'https://github.com/tryzealot/zealot');
-    return;
-  }
-
-  fetch(newVesionUrl, {
-    method: 'GET',
-    headers: {
-      'X-Requested-With': 'XMLHttpRequest',
-      'Accept': 'application/vnd.github.v3+json'
-    }
-  }).then((response) => response.json())
-    .then(json => {
-      var latest_version = json.tag_name;
-      if (compareVersion(latest_version, current_version) <= 0) { return; }
-
-      var release_link = json.html_url;
-      insert_link(elm, title + ' ' + latest_version, release_link);
-  });
-}
-
-$(document).on('turbo:load', () => {
-  check_new_version();
-});
+$(document).on("turbo:load", () => {
+  addRestartSericeEventer()
+  zealotNewVersionEventer()
+})
