+++ app/javascript/controllers/base.js
@@ -1,32 +1,27 @@
+// var _Zealot = _Zealot || {}
+
+// _Zealot.host = function () {
+//   const HOST = location.protocol + "//" + location.hostname + (location.port ? ":" + location.port : "") + "/";
+//   return HOST;
+// }
+
+// _Zealot.apperance = function () {
+//   const HOST = location.protocol + "//" + location.hostname + (location.port ? ":" + location.port : "") + "/";
+//   return HOST;
+// }
+
 const HOST = location.protocol + "//" + location.hostname + (location.port ? ":" + location.port : "") + "/";
 window.HOST = HOST;
 
 $(document).on("turbo:load", function () {
-  // fix body height for AdminLTE 2.4.0 and turbolinks 5
-  $(window).trigger("resize");
-
-  // enable tooltip global
+  // Fix tooltip toggle animation
   $("[data-toggle='tooltip']").tooltip();
 
-  // auto switch dark mode
+  // Dark mode auto enable
   var isDarkMode = window.matchMedia("(prefers-color-scheme: dark)").matches;
   if (SITE_APPEARANCE == "dark" || (SITE_APPEARANCE == "auto" && isDarkMode)) {
     $("body").addClass("dark-mode");
     $(".main-header").addClass("navbar-dark").removeClass("navbar-white");
     $(".main-sidebar").addClass("sidebar-dark-primary").removeClass("sidebar-light-primary");
   }
-
-  // // fix collapse with no response
-  // $("[data-widget="collapse"]").each(function () {
-  //   $(this).on("click", function () {
-  //     var card = $(this).parents(".card");
-  //     $(card).removeClass("collapsed-box");
-  //     $(card).boxWidget("toggle");
-  //   })
-  // });
-
-  // $(document).Toasts('create', {
-  //   title: 'Toast Title',
-  //   body: 'Lorem ipsum dolor sit amet, consetetur sadipscing elitr.'
-  // })
 });
