+++ app/javascript/controllers/debug_files/index.js
@@ -1,5 +1,4 @@
 $(document).on('turbo:load', function () {
-  alert('dddddd');
   $('.debug-file-toggle').on('click', function () {
     debug_file_id = $(this).data('id');
     $('#debug-file-metadata-' + debug_file_id).toggleClass('d-none');
