+++ app/javascript/controllers/index.js
@@ -1,6 +1,5 @@
 import "./base.js";
 import "./debug_files";
 import "./releases";
-import "./teardown/upload";
 import "./udid";
 import "./admin";
