--- app/javascript/controllers/teardown/upload.js
@@ -1,28 +1,0 @@
-// import * as FilePond from 'filepond';
-// import 'filepond/dist/filepond.min.css';
-
-
-// $(document).on('turbo:load', function () {
-// //   // Turn input element into a pond
-// const inputElement = document.querySelector('input[type="file"]');
-// const pond = FilePond.create( inputElement );
-
-//   // // Turn input element into a pond with configuration options
-//   // $('.my-pond').filepond({
-//   //     allowMultiple: true
-//   // });
-
-//   // // Set allowMultiple property to true
-//   // $('.my-pond').filepond('allowMultiple', false);
-
-//   // // Listen for addfile event
-//   // $('.my-pond').on('FilePond:addfile', function(e) {
-//   //     console.log('file added event', e);
-//   // });
-
-//   // // Manually add a file using the addfile method
-//   // $('.my-pond').filepond('addFile', 'index.html').then(function(file){
-//   //   console.log('file added', file);
-//   // });
-
-// });
