+++ app/javascript/frameworks.js
@@ -1,4 +1,9 @@
-import "./jquery"
+import jquery from "jquery"
+window.jQuery = jquery
+window.$ = jquery
+
+import clipboard from "clipboard"
+window.ClipboardJS = clipboard
+
 import "bootstrap"
 import "admin-lte"
-import "clipboard"
