--- app/javascript/jquery.js
@@ -1,3 +1,0 @@
-import jquery from "jquery"
-window.jQuery = jquery
-window.$ = jquery
