--- app/javascript/packs/admin_service.js
@@ -1,6 +1,0 @@
-import { restartService } from "javascripts/admin/service";
-
-$(document).on("turbo:load", () => {
-  var button = $("#restart-service-button");
-  button.on("click", () => restartService(button));
-});
