--- app/javascript/stylesheetsaaaa/admin/index.scss
@@ -1,4 +1,0 @@
-@import 'system_info';
-@import 'web_hooks';
-@import 'settings';
-@import 'logs';
