--- app/javascript/stylesheetsaaaa/admin/logs.scss
@@ -1,33 +1,0 @@
-.system-log {
-  background: #404040;
-
-  ol {
-    margin-left: 60px;
-    padding:10px 0;
-    margin-bottom: 0;
-    border-left: 1px solid #4c4c4c;
-    // counter-reset: number;
-    // list-style-type: none;
-    // background: #404040;
-
-    li {
-      color: #dedede;
-
-      // &::before {
-      //   counter-increment: number;
-      //   content: counter(number)"\a0";
-      // }
-
-      &:hover {
-        background-color: #2d3238;
-      }
-
-      p {
-        margin: 0;
-        color: #dedede;
-        line-height:24px;
-        padding-left: 20px;
-      }
-    }
-  }
-}
