--- app/javascript/stylesheetsaaaa/admin/settings.scss
@@ -1,77 +1,0 @@
-
-.form-group {
-  .switch {
-    position: relative;
-    display: inline-block;
-
-    .slider {
-      position: absolute;
-      cursor: pointer;
-      top: 0;
-      left: 0;
-      right: 0;
-      bottom: 0;
-      background-color: #ccc;
-      -webkit-transition: .4s;
-      transition: .4s;
-    }
-
-    .slider:before {
-      position: absolute;
-      content: "";
-      background-color: white;
-      -webkit-transition: .4s;
-      transition: .4s;
-    }
-
-    .slider.round {
-      border-radius: 34px;
-    }
-
-    .slider.round:before {
-      border-radius: 50%;
-    }
-  }
-
-  /* Hide default HTML checkbox */
-  .switch input {
-    opacity: 0;
-    width: 0;
-    height: 0;
-  }
-
-  .switch-lg {
-    width: 48px;
-    height: 24px;
-
-    .slider:before {
-      height: 14px;
-      width: 14px;
-      left: 4px;
-      bottom: 5px;
-    }
-  }
-
-  input:checked + .slider {
-    background-color: #2196F3;
-  }
-
-  input:focus + .slider {
-    box-shadow: 0 0 1px #2196F3;
-  }
-
-  input:checked + .slider:before {
-    -webkit-transform: translateX(26px);
-    -ms-transform: translateX(26px);
-    transform: translateX(26px);
-  }
-}
-
-.system-info {
-  dd {
-    .disabled {
-      background-color: #e9ecef;
-      color: gray;
-    }
-  }
-}
