--- app/javascript/stylesheetsaaaa/admin/system_info.scss
@@ -1,60 +1,0 @@
-dl.system-info {
-  padding: 10px;
-  font-size: 16px;
-  margin-bottom: 0;
-
-  dt {
-    margin-bottom: 8px;
-  }
-
-  dd {
-    margin-bottom: 20px;
-
-    pre {
-      border: 0;
-      background-color: #f7f8f9;
-      color: #2f2936;
-
-      #current-version {
-        float: none;
-      }
-
-      #new-version {
-        float: none;
-
-        a {
-          color: rgb(233, 120, 120);
-
-          i {
-            padding-right: 5px;
-          }
-        }
-      }
-
-      @media (min-width: 768px) {
-        #current-version {
-          float: left !important;
-        }
-
-        #new-version {
-          float: right !important;
-        }
-      }
-
-      @media (max-width: 767px) {
-        #current-version {
-          float: none;
-        }
-
-        #new-version {
-          float: none;
-        }
-      }
-    }
-  }
-}
-
-.dark-mode dl.system-info dd pre {
-  background-color: #2f2936;
-  color: #f7f8f9;
-}
