--- app/javascript/stylesheetsaaaa/admin/web_hooks.scss
@@ -1,10 +1,0 @@
-#webhooks-list {
-  @extend .mt-4 !optional;
-
-  .bg-light-gray {
-    margin-right: 5px;
-    background-color: #f3f2f2;
-    color: #585858;
-    text-shadow: none;
-  }
-}
