--- app/javascript/stylesheetsaaaa/app.scss
@@ -1,152 +1,0 @@
-$ios-install-issue-color: rgb(245, 39, 90);
-
-// index page
-.card-body {
-  .direct-chat-messages {
-    padding: 0;
-  }
-
-  .qr_code {
-    img {
-      width: 100%;
-    }
-  }
-}
-
-.widget-app {
-  .app-icon {
-    margin: 0 auto;
-    padding: 3px;
-    width: 100px;
-
-    max-width: 120px;
-    max-height: 120px;
-  }
-
-  .profile-username {
-    padding-top: 0.3rem;
-    font-size: 1.2em;
-  }
-
-  .created-time {
-    font-size: 0.4em;
-  }
-}
-
-.dark-mode .widget-app .app-icon {
-  background-color: #fff;
-}
-
-// show page
-.cover {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background-color: #000;
-  opacity: .7;
-  z-index: 1050;
-}
-
-.app-limited-tips {
-  position: fixed;
-  width: 100%;
-  text-align: right;
-  z-index: 1051;
-  padding-top: 14px;
-  box-sizing: border-box;
-  top: 0;
-}
-
-.app-limited-tips span {
-  text-align: left;
-  vertical-align: middle;
-  font-size: 16px;
-  color: #fff;
-  margin-right: 10px;
-  font-weight: 700;
-  line-height: 30px;
-  display: inline-block;
-}
-
-.app-limited-tips img {
-  width: 50px;
-  display: inline-block;
-  vertical-align: middle;
-  margin-bottom: 13px;
-  margin-right: 30px;
-}
-
-@media (min-width: 420px) {
-  .app-limited-tips span {
-    margin-right: 15px;
-  }
-
-  .app-limited-tips img {
-    margin-right: 20px;
-  }
-}
-
-// App 列表
-.apps {
-  .app-icon {
-    width: 25px;
-    height: 25px;
-    margin-top: -4px;
-    margin-right: 5px;
-  }
-}
-
-// App 详情
-.app-detail {
-  .app-icon {
-    width: 50px;
-    height: 50px;
-    float: left;
-  }
-
-  .app-name {
-    padding-top: 6px;
-    padding-left: 60px;
-  }
-
-  .app-identifier {
-    padding-top:4px;
-    color: #9f9f9f;
-    font-size: 0.9em;
-  }
-
-  .box-tools>.label {
-    margin-right: 3px;
-  }
-
-  .app-metadata {
-    list-style: none;
-    font-size: 1.2em;
-    padding: 0;
-
-    li {
-      display: inline-block;
-      padding: 0 10px;
-
-      i {
-        margin-right: 5px;
-      }
-    }
-  }
-
-  .app-actions {
-    .ios-install-issues {
-      padding-top: 10px;
-      font-size: 0.8em;
-      color: $ios-install-issue-color;
-
-      a {
-        color: $ios-install-issue-color;
-        text-decoration: underline;
-        padding-left: 3px;
-      }
-    }
-  }
-}
