--- app/javascript/stylesheetsaaaa/application.scss
@@ -1,10 +1,0 @@
-@import "bootstrap/scss/bootstrap.scss";
-@import "admin-lte/build/scss/adminlte.scss";
-
-@import 'layout';
-@import 'dashboard';
-@import 'app';
-@import 'channel';
-@import 'debug_files';
-@import 'devise';
-@import 'admin/index';
