--- app/javascript/stylesheetsaaaa/channel.scss
@@ -1,17 +1,0 @@
-.product-list-in-card {
-  .item {
-    .product-img {
-      margin-left: 10px;
-    }
-
-    .product-info {
-      margin-left: 70px;
-      margin-right: 10px;
-    }
-  }
-}
-
-.filter-nav {
-  // border: 1px solid #000;
-  @extend .card !optional;
-}
