--- app/javascript/stylesheetsaaaa/debug_files.scss
@@ -1,57 +1,0 @@
-.debug-file-section {
-  position: relative;
-  width: 100%;
-  cursor: pointer;
-
-  .debug-file-section-header {
-    display: block;
-    position: relative;
-
-    .debug-file-toggle {
-      padding: 10px 2px;
-    }
-
-    .debug-file-section-title {
-      display: inline-block;
-      font-size: 20px;
-      font-weight: bold;
-      margin: 0;
-      line-height: 1;
-
-      .debug-file-version, .debug-file-device-type {
-        padding-right: 5px;
-      }
-
-      .debug-file-device-type {
-        color: rgb(167, 181, 192);
-        font-size: 18px;
-      }
-    }
-
-    .debug-file-build-version {
-      color: rgb(161, 171, 187);
-      font-size: 14px;
-    }
-
-    .debug-file-section-size {
-      float: right;
-      margin-top: -50px;
-      margin-bottom: -5px;
-    }
-  }
-
-  .debug-file-section-body {
-    border-top-left-radius: 0;
-    border-top-right-radius: 0;
-    border-bottom-right-radius: 3px;
-    border-bottom-left-radius: 3px;
-
-    .table {
-      cursor: default;
-    }
-  }
-
-  .debug-file-section-body table {
-    margin-bottom: 0;
-  }
-}
