--- app/javascript/stylesheetsaaaa/devise.scss
@@ -1,25 +1,0 @@
-.social-auth-links {
-  i.icon {
-    float: right !important;
-    margin-right: 0;
-    margin-top: 0.3rem;
-  }
-
-  .button_to {
-    margin-bottom: 8px;
-  }
-}
-
-@media (min-width: 768px) {
-  .auth-providers {
-    border-left: 1px solid #e2dee6;
-  }
-}
-
-@media (max-width: 767px) {
-  .auth-providers {
-    border-top: 1px solid #e2dee6;
-    margin-top: 20px;
-    padding-top: 10px;
-  }
-}
