--- app/javascript/stylesheetsaaaa/layout.scss
@@ -1,70 +1,0 @@
-
-.navbar-empty {
-  @extend .navbar-collapse !optional;
-  @extend .justify-content-md-center !optional;
-
-  .brand-link {
-    padding: 0;
-
-    [class*="navbar-white"] & {
-      color: $black;
-    }
-
-    [class*="navbar-black"] & {
-      color: $white;
-    }
-  }
-}
-
-.layout-top-nav .wrapper .main-header .navbar-empty .brand-image {
-  margin-top: -0.1rem;
-}
-
-.card-body {
-  .table {
-    margin-bottom: 0;
-  }
-}
-
-.login-box {
-  margin: auto;
-}
-
-.btn-icon {
-  margin-right: 10px;
-
-  .fa,
-  .fab,
-  .far,
-  .fas {
-    margin-right: 5px;
-  }
-}
-
-i.icon {
-  margin-right: 5px;
-}
-
-.timeline-item {
-  .time {
-    i.icon {
-      margin-right: 3px;
-    }
-  }
-}
-
-.dark-mode {
-  .nav-item .active {
-    color: #fff;
-    background-color: #3f6791;
-  }
-
-  a {
-    color: #64adfb;
-  }
-
-  .callout {
-    border-color: #3f474e;
-    box-shadow: 0 1px 3px rgb(0 0 0 / 12%), 0 1px 2px rgb(0 0 0 / 24%);
-  }
-}
