+++ package.json
@@ -2,6 +2,7 @@
   "name": "zealot",
   "version": "4.4.1",
   "private": true,
+  "license": "MIT",
   "dependencies": {
     "@hotwired/turbo-rails": "^7.1.3",
     "@rails/actioncable": "^7.0.3",
@@ -10,7 +11,6 @@
     "clipboard": "^2.0.11",
     "compare-versions": "^4.1.3",
     "esbuild": "^0.14.50",
-    "filepond": "^4.30.4",
     "rails-erb-loader": "^5.5.2",
     "sass": "^1.54.0"
   },
